+++ CHANGELOG.md
@@ -5,7 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-[Unreleased](https://github.com/jshwi/hin/compare/v0.1.4...HEAD)
+[Unreleased](https://github.com/jshwi/hin/compare/v0.1.5...HEAD)
+------------------------------------------------------------------------
+
+[0.1.5](https://github.com/jshwi/hin/releases/tag/v0.1.5) - 2023-06-17
 ------------------------------------------------------------------------
 ### Fixed
 - allow committing child when not the only file